sell

+++ TowerDefense/src/core/applicationService/warriorServices/TowerMarket.java
@@ -4,8 +4,8 @@ import core.contract.DefenderConstants;
 import core.domain.warriors.defenders.towers.Tower;
 
 public class TowerMarket {
-	public <T extends Tower> double sellTower(T tower){
-		double amount = 0;
+	public <T extends Tower> long sellTower(T tower){
+		long amount = 0;
 		try {
 				amount = (tower.cost() * DefenderConstants.Sell_Percentage)/100;
 		} catch (Exception e) {
